Allow passing in kubeconfig contents during runtime
e.g. for receptorctl, --param secret_kube_config="apiVersion: v1..."

if the value for --param starts with @, it is interpreted as a file. In this case the file is read and the contents are passed to receptor

e.g. --param secret_kube_config=@/home/sfoster/.kube/config

Note: the @ indicating a filename works for any --param, not just secrete_kube_config