Let currently ignored errors panic when they arrise
There are currently several locations where errors are not checked.
This commit checks if they are non nil and panics when so.

At some occasions it would have been possible to just dump the new
error into an existing return but I assume this would lead to more
subtle and possible missed errors, so I went with a violent program
explosion instead