Add cobra and viper as cmdline parser.
Currently the cmdlineparser `cmdline` is used for bootstrapping
receptor instances. With the current use of it, receptor is a
singleton and can't be created multiple times per process, which
makes testing complicated. this PR adds cobra and viper as cmdline
parsers. Both are rather omnipresent in the go world and bring
multiple advantages like sourcing configuration options from
different locations and extra features.

In this commit I added a new main package alongside the old one
which uses said tools and a new struct set. It does not break
existing functionality. My goal is to offer this feature side by
side to see how the reception is and to iron out bugs.
Should the old solution be deprecated I expect less complicated
bootstrapping code that allows for better testing  and higher
robustness.