Sign and verify work submissions
Remote work submissions can be digitally signed by the sender. The
target node will verify the signature of the work command before
starting the work unit.

A pair of RSA public and private keys are created offline and
distributed to the nodes. The public key should be on the node receiving
work (PKIX format). The private key should be on the node submitting
work (PKCS1 format).