Revert cobra and viper addition.
My goal was to phase out cmdline and replace it with cobra and viper.
Since then, cb0dfd4b982234e53794fe47cc11535b5155747b, there was no
progress with that and it remains unused (is broken in main branch).
Therefore I reverted relevant commits to prevent this from just being
dead weight in the repo.