Give receptor subprocess a nodeID
When work units are started, a separate receptor process is ran
If this receptor process does not have a node ID, receptor will try to
use os.hostname. If hostname begins with "localhost", receptor will
error out. To avoid this scenario, we should just give the receptor a
nodeID. Since this receptor process is not joining a mesh, the name
really doesn't matter.