Fix race condition in TestTCPProxyServiceOutbound
All tests used a shared netceptor.Listener pointer. Because the
function being tested uses go routines this shared listener would
sometimes cause one test to signal the AcceptChan of another
test. When this happened to the "Fail to accept input connections"
test, it never expects to call the Dial function. But it does because
it received the AcceptResult from another test.