Remove test results uploading in coverage reporting
Removes the uploading of the /tmp/receptor-testing directory as a repo
artifact from the coverage reporting workflow. When running `make
coverage` the functional tests are not run, and it is these tests that
create that directory. The functional tests are run in the pull
request workflow, and the test results are uploaded, so that isn't
being lost.